UPDATE: Updated the file structure for the video API
