+++ supervision/__init__.py
@@ -130,11 +130,17 @@ from supervision.utils.image import (
 from supervision.utils.notebook import plot_image, plot_images_grid
 from supervision.utils.video import (
     FPSMonitor,
-    Video,
-    VideoInfo,
     VideoSink,
     get_video_frames_generator,
     process_video,
+)
+
+from supervision.video.utils import (
+    VideoInfo
+)
+
+from supervision.video.core import (
+    Video
 )
 
 __all__ = [
