+++ supervision/video/utils.py
@@ -1,0 +1,58 @@
+from dataclasses import dataclass
+from enum import Enum
+import cv2
+
+class SOURCE_TYPE(Enum):
+    VIDEO_FILE = "VIDEO_FILE"
+    WEBCAM = "WEBCAM"
+    RTSP = "RTSP"
+
+@dataclass
+class VideoInfo:
+    """
+    A class to store video information, including width, height, fps and
+        total number of frames.
+
+    Attributes:
+        width (int): width of the video in pixels
+        height (int): height of the video in pixels
+        fps (int): frames per second of the video
+        total_frames (Optional[int]): total number of frames in the video,
+            default is None
+
+    Examples:
+        ```python
+        import supervision as sv
+
+        video_info = sv.VideoInfo.from_video_path(video_path=<SOURCE_VIDEO_FILE>)
+
+        video_info
+        # VideoInfo(width=3840, height=2160, fps=25, total_frames=538)
+
+        video_info.resolution_wh
+        # (3840, 2160)
+        ```
+    """
+
+    width: int
+    height: int
+    fps: int
+    total_frames: int | None = None
+    source_type: SOURCE_TYPE | None = None
+
+    @classmethod
+    def from_video_path(cls, video_path: str) -> "VideoInfo":
+        video = cv2.VideoCapture(video_path)
+        if not video.isOpened():
+            raise Exception(f"Could not open video at {video_path}")
+
+        width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
+        height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
+        fps = round(video.get(cv2.CAP_PROP_FPS))
+        total_frames = int(video.get(cv2.CAP_PROP_FRAME_COUNT))
+        video.release()
+        return VideoInfo(width, height, fps, total_frames)
+
+    @property
+    def resolution_wh(self) -> tuple[int, int]:
+        return self.width, self.height
